ai_core/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.51.0
httpx==0.27.2
SQLAlchemy==2.0.32
alembic==1.13.2
numpy==1.26.4
apscheduler==3.10.4
orjson==3.10.7